Moving Snapshot into its own module...
