+++ src/db.rs
@@ -18,8 +18,8 @@ use ffi_util::opt_bytes_to_ptr;
 use util::to_cpath;
 
 use crate::{
-    handle::Handle, ColumnFamily, ColumnFamilyDescriptor, DBIterator, DBRawIterator, DBVector, Direction, Error,
-    IteratorMode, Options, ReadOptions, WriteOptions, ops::{self, *},
+    handle::Handle, ColumnFamily, ColumnFamilyDescriptor, DBIterator, DBRawIterator, Direction, Error,
+    IteratorMode, Options, ReadOptions, WriteOptions, Snapshot, ops,
 };
 
 use libc::{self, c_char, c_int, c_void, size_t};
@@ -83,131 +83,7 @@ pub struct WriteBatch {
     pub(crate) inner: *mut ffi::rocksdb_writebatch_t,
 }
 
-/// A consistent view of the database at the point of creation.
-///
-/// ```
-/// use rocksdb::{DB, IteratorMode, Options};
-/// # use rocksdb::TemporaryDBPath;
-///
-/// let path = "_path_for_rocksdb_storage3";
-/// # let path = TemporaryDBPath::new();
-/// # {
-///
-///     let db = DB::open_default(&path).unwrap();
-///     let snapshot = db.snapshot(); // Creates a longer-term snapshot of the DB, but closed when goes out of scope
-///     let mut iter = snapshot.iterator(IteratorMode::Start); // Make as many iterators as you'd like from one snapshot
 
-/// # }
-/// ```
-///
-pub struct Snapshot<'a> {
-    db: &'a DB,
-    pub(crate) inner: *const ffi::rocksdb_snapshot_t,
-}
-
-impl<'a> Snapshot<'a> {
-    pub fn new(db: &DB) -> Snapshot {
-        let snapshot = unsafe { ffi::rocksdb_create_snapshot(db.inner) };
-        Snapshot {
-            db,
-            inner: snapshot,
-        }
-    }
-
-    pub fn iterator(&self, mode: IteratorMode) -> DBIterator {
-        let readopts = ReadOptions::default();
-        self.iterator_opt(mode, readopts)
-    }
-
-    pub fn iterator_cf(
-        &self,
-        cf_handle: ColumnFamily,
-        mode: IteratorMode,
-    ) -> Result<DBIterator, Error> {
-        let readopts = ReadOptions::default();
-        self.iterator_cf_opt(cf_handle, readopts, mode)
-    }
-
-    pub fn iterator_opt(&self, mode: IteratorMode, mut readopts: ReadOptions) -> DBIterator {
-        readopts.set_snapshot(self);
-        DBIterator::new(self.db, &readopts, mode)
-    }
-
-    pub fn iterator_cf_opt(
-        &self,
-        cf_handle: ColumnFamily,
-        mut readopts: ReadOptions,
-        mode: IteratorMode,
-    ) -> Result<DBIterator, Error> {
-        readopts.set_snapshot(self);
-        DBIterator::new_cf(self.db, cf_handle, &readopts, mode)
-    }
-
-    pub fn raw_iterator(&self) -> DBRawIterator {
-        let readopts = ReadOptions::default();
-        self.raw_iterator_opt(readopts)
-    }
-
-    pub fn raw_iterator_cf(&self, cf_handle: ColumnFamily) -> Result<DBRawIterator, Error> {
-        let readopts = ReadOptions::default();
-        self.raw_iterator_cf_opt(cf_handle, readopts)
-    }
-
-    pub fn raw_iterator_opt(&self, mut readopts: ReadOptions) -> DBRawIterator {
-        readopts.set_snapshot(self);
-        DBRawIterator::new(self.db, &readopts)
-    }
-
-    pub fn raw_iterator_cf_opt(
-        &self,
-        cf_handle: ColumnFamily,
-        mut readopts: ReadOptions,
-    ) -> Result<DBRawIterator, Error> {
-        readopts.set_snapshot(self);
-        DBRawIterator::new_cf(self.db, cf_handle, &readopts)
-    }
-
-    pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<DBVector>, Error> {
-        let readopts = ReadOptions::default();
-        self.get_opt(key, readopts)
-    }
-
-    pub fn get_cf<K: AsRef<[u8]>>(
-        &self,
-        cf: ColumnFamily,
-        key: K,
-    ) -> Result<Option<DBVector>, Error> {
-        let readopts = ReadOptions::default();
-        self.get_cf_opt(cf, key.as_ref(), readopts)
-    }
-
-    pub fn get_opt<K: AsRef<[u8]>>(
-        &self,
-        key: K,
-        mut readopts: ReadOptions,
-    ) -> Result<Option<DBVector>, Error> {
-        readopts.set_snapshot(self);
-        self.db.get_opt(key.as_ref(), &readopts)
-    }
-
-    pub fn get_cf_opt<K: AsRef<[u8]>>(
-        &self,
-        cf: ColumnFamily,
-        key: K,
-        mut readopts: ReadOptions,
-    ) -> Result<Option<DBVector>, Error> {
-        readopts.set_snapshot(self);
-        self.db.get_cf_opt(cf, key.as_ref(), &readopts)
-    }
-}
-
-impl<'a> Drop for Snapshot<'a> {
-    fn drop(&mut self) {
-        unsafe {
-            ffi::rocksdb_release_snapshot(self.db.inner, self.inner);
-        }
-    }
-}
 
 impl DB {
     /// Open a database with default options.
@@ -1188,6 +1064,8 @@ impl<'a> DBPinnableSlice<'a> {
 #[test]
 fn test_db_vector() {
     use std::mem;
+    use crate::prelude::*;
+
     let len: size_t = 4;
     let data = unsafe { libc::calloc(len, mem::size_of::<u8>()) as *mut u8 };
     let v = unsafe { DBVector::from_c(data, len) };
@@ -1197,7 +1075,7 @@ fn test_db_vector() {
 
 #[test]
 fn external() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
@@ -1213,7 +1091,7 @@ fn external() {
 
 #[test]
 fn errors_do_stuff() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
@@ -1233,7 +1111,7 @@ fn errors_do_stuff() {
 
 #[test]
 fn writebatch_works() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
@@ -1276,7 +1154,7 @@ fn writebatch_works() {
 
 #[test]
 fn iterator_test() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
@@ -1300,7 +1178,7 @@ fn iterator_test() {
 
 #[test]
 fn snapshot_test() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
@@ -1322,7 +1200,7 @@ fn snapshot_test() {
 
 #[test]
 fn set_option_test() {
-    use crate::TemporaryDBPath;
+    use crate::{prelude::*, TemporaryDBPath};
 
     let path = TemporaryDBPath::new();
     {
